none of this works yet. trying to make ratings update on splash
