+++ routes/reviews.js
@@ -4,7 +4,8 @@ var router = express.Router();
 const { Haunt, Review } = require('../db/models');
 const { asyncHandler, csrfProtection } = require("./utils");
 const { check, validationResult } = require("express-validator");
-const db = require("../db/models");
+const { averageScore } = require('../auth.js');
+
 
 
 const reviewValidator = [
@@ -18,33 +19,6 @@ const reviewValidator = [
       .withMessage("A rating is required"),
 ];
 
-// Helper function to calculate and average review average scores
-async function averageScore(hauntId) {
-    // parameter is integer hauntId (Review.hauntId or Haunt.id)
-    // returns void - should just update the Haunt instance
-
-    const hauntToUpdate = await Haunt.findByPk(hauntId);
-
-    const hauntReviews = await Review.findAll({
-        where: { hauntId }
-    });
-
-    const numOfReviews = hauntReviews.length;
-
-    const reviewSum = hauntReviews.reduce((accl, rating, i) => {
-        rating = hauntReviews[i].score;
-        return accl + rating;
-    }, 0);
-
-    const reviewAvg = +parseFloat(reviewSum / numOfReviews).toFixed(2);
-
-    const updatedHaunt = await hauntToUpdate.update({
-        score: reviewAvg
-    });
-
-    await updatedHaunt.save();
-    return;
-}
 
 
 // GET page with form for new review
